pyproject.toml
[tool.poetry]
name = "speech-service"
version = "0.1.0"
description = "Diffaction speech-to-text service"
authors = ["artemowkin <aloskan@sdvor.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai-whisper = "^20231117"
boto3 = "^1.34.113"
aio-pika = "^9.4.1"
pydantic = "^2.7.3"
pydantic-settings = "^2.3.1"
sentry-sdk = "^2.5.0"
pyright = "^1.1.366"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
